refactor(sidebar): split sidebar root into menu/resize/sections submodules
Splits src/skills/default-layout/sidebar/index.tsx (876 LOC) into a
thin render shell plus four focused submodules. The Sidebar export and
every re-export (filterItems, providerOf, ItemRow, SearchableSidebarSection
+ types) stays identical so callers don't move.

- menuItems.ts: pure data — SidebarContextMenuState / SidebarMenuHandlers
  types, extensionToggleState, buildSidebarMenuItems.
- contextMenu.tsx: useSidebarContextMenu hook owning the menu open/close
  state plus every project/worktree/session/extension handler, the kind
  classifier, and the rename/delete prompts.
- resize.ts: useSidebarResize hook owning asideRef + the mousemove/up
  drag loop with MIN/MAX width constants.
- sections.ts: pure composeSidebarSections + resolveSidebarItems —
  extension auto-injection (project/user bucketing) and extra-section
  $ref resolution.
- index.tsx: thin shell composing the hooks, plus a SidebarSectionBlock
  helper for the non-searchable section render (so the map body stays at
  one screen).

Behavior unchanged. Public surface unchanged.
